+++ Embrace.API/Infrastructure/Persistence/Doacao.cs
@@ -2,12 +2,12 @@
 {
     public class Doacao
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Tipo { get; set; } // Roupas, Alimentos, etc.
         public int Quantidade { get; set; }
         public DateTime DataRecebida { get; set; }
 
-        public int AcaoSolidariaId { get; set; }
+        public long AcaoSolidariaId { get; set; }
         public AcaoSolidaria AcaoSolidaria { get; set; }
     }
 }
