+++ Embrace.API/Infrastructure/Persistence/Meta.cs
@@ -2,10 +2,10 @@
 {
     public class Meta
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Descricao { get; set; }
         public int QuantidadeEsperada { get; set; }
-        public int AcaoSolidariaId { get; set; }
+        public long AcaoSolidariaId { get; set; }
 
         public AcaoSolidaria AcaoSolidaria { get; set; }
     }
